Fix obj build compile error
change ~0 to (ULONG)~0

+++ open-vm-tools/lib/include/vm_basic_defs.h
@@ -691,7 +691,7 @@ typedef int pid_t;
 #ifndef USES_OLD_WINDDK
 #if defined(VMX86_LOG)
 #ifdef _WIN64
-#define WinDrvPrint(arg, ...) DbgPrintEx(DPFLTR_IHVDRIVER_ID, ~0, arg, __VA_ARGS__)
+#define WinDrvPrint(arg, ...) DbgPrintEx(DPFLTR_IHVDRIVER_ID, (ULONG)~0, arg, __VA_ARGS__)
 #else
 #define WinDrvPrint(arg, ...) DbgPrint(arg, __VA_ARGS__)
 #endif
